refactor(components): organize navigation components into a dedicated folder
