+++ src/app/settings/page.js
@@ -1,6 +1,6 @@
 import React from "react";
 import PrivateRoute from "@/components/routes/PrivateRoute";
-import Sidebar from "@/components/Sidebar";
+import Sidebar from "@/components/navigation/Sidebar";
 
 export default function page() {
   return (
