+++ src/components/navigation/Drawer.js
@@ -6,7 +6,7 @@ import { IoMdSettings } from "react-icons/io";
 import Image from "next/image";
 import Link from "next/link";
 import { useAuth } from "@/context/AuthContext";
-import CreatePost from "./post/CreatePost";
+import CreatePost from "../post/CreatePost";
 
 const Drawer = () => {
   const { handleLogout, currentUser } = useAuth();
