+++ src/components/post/PostDetails.js
@@ -7,9 +7,9 @@ import { FaRegClock } from "react-icons/fa6";
 import { FaRegCalendar } from "react-icons/fa";
 import { format } from "date-fns";
 import { useMessage } from "@/context/MessageContext";
-import ConfirmationModal from "../post/ConfirmationModal";
-import LikeButton from "../post/LikeButton";
-import Menu from "../post/Menu";
+import ConfirmationModal from "./ConfirmationModal";
+import LikeButton from "./LikeButton";
+import Menu from "./Menu";
 import Category from "./Category";
 import Tag from "./Tag";
 import api from "@/utils/axiosInstance";
